Extract Lazy operatio to own file
